Apply more suggestions

+++ src/chatdDb.h
@@ -215,7 +215,6 @@ public:
             {
                 Buffer refs;
                 stmt.blobCol(9, refs);
-                // todo review this one and adjust
                 refs.read(0, msg->backRefs);
             }
 
@@ -233,7 +232,7 @@ public:
                 chatd::MsgCommand *msgCmd = new chatd::MsgCommand(opcode, mChat.chatId(), userid, msgid, ts, updated, keyid);
                 Buffer buf;
                 stmt.blobCol(11, buf);
-                msgCmd->setMsg(buf.buf(), static_cast<uint32_t>(buf.dataSize()));
+                msgCmd->setMsg(buf.buf(), buf.dataSize());
 
                 queue.back().msgCmd = msgCmd;
             }
